[IMP] school_base: Changed m2m to o2m
