+++ adm/static/src/css/family_select.css
@@ -7,6 +7,8 @@
   flex: 0 1 10%;
   min-width: 120px;
   max-width: 120px;
+  display: flex;
+  flex-direction: column;
 }
 .o_adm_family_select_item img {
   opacity: 0.5;
@@ -19,6 +21,18 @@
 }
 .o_adm_family_selection li {
   list-style: none;
+}
+
+#parent_relationship_card_list .o_adm_hide_if_parent {
+  display: none;
+}
+
+#sibling_relationship_card_list .o_adm_hide_if_sibling {
+  display: none;
+}
+
+#other_relationship_card_list .o_adm_hide_if_other {
+  display: none;
 }
 
 /*# sourceMappingURL=family_select.css.map */
