+++ adm/static/src/js/inquiry/inquiry.js
@@ -1,47 +1,24 @@
-odoo.define('adm.inquiry', require => {
+odoo.define('adm.application.student_info', require => {
     "use strict"
-    // TODO: Make this script work with Widgets
 
-    // We just wait to some things to be done before using JQuery :)
-    // This can be unneeded if we drop JQuery in this module (I hope some day...)
     require('web.core');
 
-    let studentCount = 1;
+    var studentCount = 1;
 
-    function showGradeLevelsBySchoolCode(event) {
-        const $currentTarget = $(event.currentTarget);
-        const $selectGradeLevels = $currentTarget.parents('[role="tabpanel"]').find('[name="selStudentInterestGradeLevel"]');
-        const schoolCodeId = $currentTarget.val();
-        
-        $selectGradeLevels.children("option:gt(0)").hide();
-        $selectGradeLevels.children("option[data-school-code='" + schoolCodeId + "']").show();
-
-        if ($selectGradeLevels.children("option:selected").is(":hidden")) {
-            $selectGradeLevels.children("option:nth(0)").prop("selected", true);
-        }
-    }
-    
-    function recomputaTabStudent() {
-        $('#studentsNavbar').find('li').each((i, el) => {
-            const $el = $(el);
-            $el.children('a.student').html('Student '+(i+1)+'');
-        });
-        $('#studentsNavbar').find('li > a.student').last().click();
-        $('#studentsCount').val($('#studentsNavbar').find('li > a').length-1);
-    }
     function removeStudent(idStudent) {
+        studentCount--;
         $(`#navStudent${idStudent}`).remove();
         $(`#student${idStudent}`).remove();
-        recomputaTabStudent();
     }
 
     function addStudent() {
         studentCount++;
         var htmlTab =
             `<li class="nav-item" style="position: relative" id="navStudent${studentCount}">
-        <a class="nav-link student" id="student${studentCount}-tab" data-toggle="tab" href="#student${studentCount}"
+        <a class="nav-link" id="student${studentCount}-tab" data-toggle="tab" href="#student${studentCount}"
             role="tab" aria-controls="student${studentCount}" aria-selected="false">Student ${studentCount}</a>
-            <i data-tabRemove="${studentCount}" class="fa fa-times remove-tab" style="position: absolute; top: -0.5em; right: 0.1em; font-size: 1.4em; color: orangered; cursor: pointer;""></i>
+            <i class="fa fa-times" style="position: absolute; top: -0.5em; right: 0.1em; font-size: 1.4em; color: orangered; cursor: pointer;"
+            onclick="removeStudent(${studentCount})"></i>
     </li>`;
 
 
@@ -60,72 +37,76 @@ odoo.define('adm.inquiry', require => {
         $(studentClonnable).removeClass("active");
         $(studentClonnable).removeClass("show");
 
-        $(studentClonnable).find('input[id]').each((i, el) => {
-            const $el = $(el);
-            const newId = el.id + '-' + studentCount;
-            $(studentClonnable).find(`label[for=${el.id}]`).attr('for', newId);
-            el.id = newId
-        });
-//        $(studentClonnable).find("#txtStudent1FirstName").attr("id", "txtStudent" + studentCount + "FirstName")
-//        $(studentClonnable).find("#txtStudent1MiddleName").attr("id", "txtStudent" + studentCount + "MiddleName")
-//        $(studentClonnable).find("#txtStudent1LastName").attr("id", "txtStudent" + studentCount + "LastName")
-//        $(studentClonnable).find("#txtStudent1Birthday").attr("id", "txtStudent" + studentCount + "Birthday")
-//        $(studentClonnable).find("#selStudent1Gender").attr("id", "selStudent" + studentCount + "Gender")
-//        $(studentClonnable).find("#selStudent1Nativelanguage").attr("id", "selStudent" + studentCount + "Nativelanguage")
-//        $(studentClonnable).find("#selStudent1SchoolYear").attr("id", "selStudent" + studentCount + "SchoolYear")
-//        $(studentClonnable).find("#selStudent1GradeLevel").attr("id", "selStudent" + studentCount + "GradeLevel")
-//        $(studentClonnable).find("#selStudent1CurrentGradeLevel").attr("id", "selStudent" + studentCount + "CurrentGradeLevel")
-//        $(studentClonnable).find("#txtStudent1CurrentSchool").attr("id", "txtStudent" + studentCount + "CurrentSchool")
-//        $(studentClonnable).find("#txtStudent1FromEnglishSchool").attr("id", "txtStudent" + studentCount + "FromEnglishSchool")
-//        $(studentClonnable).find("#txtStudent1FromEnglishSchool").attr("label", "txtStudent" + studentCount + "FromEnglishSchool")
-//        $(studentClonnable).find(".txtStudent1ExtraServices").attr("name", "txtStudent" + studentCount + "ExtraServices")
-//        $(studentClonnable).find(".txtStudent1ExtraServices").attr("id", "txtStudent" + studentCount + "ExtraServices")
-//        $(studentClonnable).find("#fileStudent1Photo").attr("name", "fileStudent" + studentCount + "Photo")
-//        $(studentClonnable).find("#fileStudent1BirthCert").attr("name", "fileStudent" + studentCount + "BirthCert")
-//        $(studentClonnable).find("#fileStudent1ReportCard").attr("name", "fileStudent" + studentCount + "ReportCard")
-//        $(studentClonnable).find("#fileStudent1ImmunizationRecord").attr("name", "fileStudent" + studentCount + "ImmunizationRecord")
-
         $(studentClonnable).find("input").each(function () {
-            var input_type = $(this).attr("type")
-            if (input_type == "checkbox") {
-                this.checked = false
-            } else {
-                $(this).val("");
+            $(this).val("");
+        });
+
+        // Grade Level List
+        var optionsGradeLevel = $('select#selStudent1CurrentGradeLevel option').clone();
+        $('#selStudent' + studentCount + 'CurrentGradeLevel').append(optionsGradeLevel);
+
+        //   document.querySelectorAll(".selectSchoolCode").forEach(function(element){
+        //         element.addEventListener("change", function(){
+        //             var schoolCodeID = $(this).find("option:selected").data("id");
+        //             $(this).parent().parent().children().last().find("option").each(function(element){
+        //                 var school_option_id = $(this).data("schoolid");
+        //                 if(schoolCodeID != -1 && schoolCodeID == school_option_id)
+        //                     $(this).removeClass('d-none')
+        //                 else
+        //                     $(this).addClass('d-none')
+        //             })
+        //             $(this).parent().parent().children().last().find("select").val(-1);
+        //         })
+        //     });
+    }
+
+    function getStates() {
+        $('#selState').html("<option value='-1'>-Select a state-</option>");
+        $.ajax({
+            url: '/admission/states',
+            type: 'GET',
+            data: {'country_id': $('#selCountry').val()},
+            success: function (data) {
+                $.each(JSON.parse(data), function (i, state) {
+                    // console.log(`<option
+                    // value='${state.id}'>${state.name}</option>`);
+                    $('#selState').append(`<option value='${state.id}'>${state.name}</option>`)
+                })
+            },
+            error: function () {
+                console.error("Un error ha ocurrido al cargar los states");
             }
         });
 
-
-        document.querySelectorAll(".selectSchoolYear").forEach(function (element) {
-            element.addEventListener("change", function () {
-                var schoolCodeID = $(this).find("option:selected").data("school_code");
-                $(this).parent().next().find("select option").each(function (element) {
-                    var school_option_id = $(this).data("school_code");
-                    if (schoolCodeID != -1 && schoolCodeID == school_option_id)
-                        $(this).removeClass('d-none')
-                    else
-                        $(this).addClass('d-none')
-                })
-                $(this).parent().next().find("select").val(-1);
-            })
-        });
-        recomputaTabStudent();
-        $(studentClonnable).find('select[name="selStudentSchoolCode"]').on('click', showGradeLevelsBySchoolCode);
     }
 
-    function showOnlyCountrysStates() {
-        // You might ask, why don't use # selector instead of getElementById
-        // The reason for that is getElementById uses native code, so is much faster
-        // And passing to jquery an element, it doesn't try to find it.
-        // So... $(getElementById) has better performance and allow us to use JQuery :)
+    function changeState() {
+        var select_state = $('#selState')
+        select_state.children("option:gt(0)").hide();
+        select_state.children("option[data-country='" + $(this).val() + "']").show();
 
-        const selectState = $(document.getElementById('selState'));
-        const selectCountry = document.getElementById('selCountry');
+        if (select_state.children("option:selected").is(":hidden")) {
+            select_state.children("option:nth(0)").prop("selected", true);
+        }
+    }
 
-        selectState.children("option:gt(0)").hide();
-        selectState.children("option[data-country='" + selectCountry.value + "']").show();
+    function ready(fn) {
+        if (document.readyState != 'loading') {
+            fn();
+        } else {
+            document.addEventListener('DOMContentLoaded', fn);
+        }
+    }
 
-        if (selectState.children("option:selected").is(":hidden")) {
-            selectState.children("option:nth(0)").prop("selected", true);
+    function toggleSecondParent() {
+        $("#section_parent_2").toggle();
+
+        if ($("#section_parent_2").css("display") == "none") {
+            $("#section_parent_2 input").attr("disabled", true);
+            $("#section_parent_2 select").attr("disabled", true);
+        } else {
+            $("#section_parent_2 input").attr("disabled", false);
+            $("#section_parent_2 select").attr("disabled", false);
         }
     }
 
@@ -151,51 +132,65 @@ odoo.define('adm.inquiry', require => {
 
     }
 
-    function toggleSecondParent() {
-        $("#section_parent_2").toggle();
+    function checkDuplicateContact() {
+        var first_name = $(this).parent().parent().parent().find(".firstname");
+        var last_name = $(this).parent().parent().parent().find(".lastname");
+        var email = $(this).parent().parent().parent().find(".email");
+        var cellphone = $(this).parent().parent().parent().find(".phone");
 
-        if ($("#section_parent_2").css("display") == "none") {
-            $("#section_parent_2 input").attr("disabled", true);
-            $("#section_parent_2 select").attr("disabled", true);
-        } else {
-            $("#section_parent_2 input").attr("disabled", false);
-            $("#section_parent_2 select").attr("disabled", false);
+        if (first_name.val() != undefined && first_name.val() != '' &&
+            last_name.val() != undefined && last_name.val() != '' &&
+            email.val() != undefined && email.val() != '' &&
+            cellphone.val() != undefined && cellphone.val() != '') {
+
+            $.ajax({
+                url: "/admission/checkDuplicateContact",
+                type: "POST",
+                data: {
+                    "firstname": first_name.val(),
+                    "lastname": last_name.val(),
+                    "email": email.val(),
+                    "cellphone": cellphone.val()
+                },
+                success: function (data) {
+                    var jsonData = JSON.parse(data);
+                    //var element_event = elem;
+                    if (jsonData.parent_name || jsonData.email || jsonData.cellphone) {
+                        $("#emailChecked").text(jsonData.email);
+                        $("#modalError").modal('show');
+                        var messageResponse = '';
+                        //element_event.val('');
+
+                        messageResponse = 'Hi, ' + first_name.val() + ' ' + last_name.val() + '<br>'
+
+                        if (jsonData.parent_name)
+                            messageResponse += 'In the system exists a contact with same name.<br>'
+
+                        if (jsonData.email)
+                            messageResponse += 'The email <b>' + email.val() + '</b> exists in the system.<br>'
+
+                        if (jsonData.cellphone)
+                            messageResponse += 'The mobile <b>' + cellphone.val() + '</b> exists in the system.<br>'
+
+                        messageResponse += "If you want to add this person to the family. Please contact with the admission support to <b>support@iae.edu</b>."
+
+                        $("#bodyMessageIssue").html(messageResponse);
+                    }
+
+                }
+            })
         }
     }
 
-    function invoiceAddress(valor) {
-        var status = valor.data
-        if(status == 1){
-            $("input[name='txtInvoiceAddress_1']").prop("checked", true);
-            $("input[name='txtInvoiceAddress_2']").prop("checked", false);
-        }else{
-            $("input[name='txtInvoiceAddress_1']").prop("checked", false);
-            $("input[name='txtInvoiceAddress_2']").prop("checked", true);
-        }
+    ready(function () {
+        document.querySelector("#add-tab").addEventListener("click", addStudent)
+        document.querySelectorAll(".checkDuplicated").forEach(
+            function (element) {
+                element.addEventListener("change", checkDuplicateContact)
+            }
+        );
 
-    }
-
-    $(document).ready(function () {
-
-        // Format Init data :P
-        showOnlyCountrysStates();
-
-        // Event Handlers
-        //document.getElementById("add-tab").addEventListener("click", addStudent);
-        document.getElementById("selCountry").addEventListener("change", showOnlyCountrysStates);
-
-        document.getElementById('showSecondParent').addEventListener('click', toggleSecondParent);
-
-        $('#checkbox_family_id').on('click',hideDataParents)
-        $('#txtInvoiceAddress_1').click(1,invoiceAddress)
-        $('#txtInvoiceAddress_2').click(2,invoiceAddress)
-
-        $('#add-tab').on('click',addStudent)
-
-        $(document).on('click', '.remove-tab', function () {
-            removeStudent($(this).attr('data-tabRemove'))
-        });
-
+        //document.querySelector("#selCountry").addEventListener("change", changeState)
 
         document.querySelectorAll(".custom-file-input").forEach(function (element) {
             element.addEventListener("change", function () {
@@ -203,67 +198,44 @@ odoo.define('adm.inquiry', require => {
                 $(this).next("label").text(fileName);
             })
         })
-        document.querySelectorAll(".selectSchoolYear").forEach(function (element) {
+
+        var optionValues = [];
+        document.querySelectorAll(".selectSchoolCode option").forEach(function (element) {
+            if (optionValues.indexOf(element.value) > -1) {
+                element.remove();
+            } else {
+                optionValues.push(element.value);
+            }
+        })
+        document.querySelectorAll(".selectSchoolCode").forEach(function (element) {
             element.addEventListener("change", function () {
-                var schoolCodeID = $(this).find("option:selected").data("school_code");
-                $(this).parent().next().find("select option").each(function (element) {
-                    var school_option_id = $(this).data("school_code");
+                var schoolCodeID = $(this).find("option:selected").data("id");
+                $(this).parent().parent().children().last().find("option").each(function (element) {
+                    var school_option_id = $(this).data("schoolid");
                     if (schoolCodeID != -1 && schoolCodeID == school_option_id)
                         $(this).removeClass('d-none')
                     else
                         $(this).addClass('d-none')
                 })
-                $(this).parent().next().find("select").val(-1);
+                $(this).parent().parent().children().last().find("select").val(-1);
             })
         });
-//        document.querySelectorAll(".selStudentSchoolCode").forEach(function (element) {
-//            element.addEventListener("change", function () {
-//                var schoolCodeID = $(this).find("option:selected").data("school_code");
-//                $(this).parent().next().find("select option").each(function (element) {
-//                    var school_option_id = $(this).data("school_code");
-//                    if (schoolCodeID != -1 && schoolCodeID == school_option_id)
-//                        $(this).removeClass('d-none')
-//                    else
-//                        $(this).addClass('d-none')
-//                })
-//                $(this).parent().next().find("select").val(-1);
-//            })
-//        });
-        document.querySelectorAll("#selSource").forEach(function (element) {
+
+        document.querySelectorAll("#inquiry_country_id").forEach(function (element) {
             element.addEventListener("change", function () {
-                if ($(this).find("option:selected").data("other") != undefined)
-                    $("#contOtherSource").removeClass("d-none")
+                var countryCode = $(this).find("option:selected").val();
+                $(this).parent().parent().children().last().find("option").each(function (element) {
+                    var country_option_id = $(this).data("country");
+                    if (countryCode != -1 && countryCode == country_option_id)
+                        $(this).removeClass('d-none')
+                    else
+                        $(this).addClass('d-none')
+                })
+                if ($("#inquiry_state_id option:not(.d-none)").length > 0)
+                    $("#inquiry_state_id option:not(.d-none)").first().attr("selected", "selected");
                 else
-                    $("#contOtherSource").addClass("d-none")
+                    $("#inquiry_state_id").val(-1);
             })
         });
-        $('select[name="selStudentSchoolCode"]').on('change', showGradeLevelsBySchoolCode).trigger('change');
-        console.log('asdasd');
     });
-
-    // Get user location
-    $.get('/adm/geolocation').then(result => {
-        if (!_.isEmpty(result)) {
-            // Country
-            const $selCountry = $('[name^=selCountry]');
-            $selCountry.each((i, selCountry) => {
-                const countryOption = _.filter(selCountry.options, option => option.dataset.code === result.country_code)
-                if (countryOption.length) {
-                    selCountry.value = countryOption[0].value;
-
-                    // State
-                    const $selState = $('[name^=selState]');
-                    $selState.each((i, selState) => {
-                        
-                        const stateOption = _.filter(selCountry.options, option => option.dataset.country === countryOption[0].value
-                                                     && option.dataset.code === result.region)
-                        if (stateOption.length) {
-                            selState.value = stateOption[0].value;
-                        }
-                    });
-                }
-            });
-            showOnlyCountrysStates();
-        }
-    })
-})
+});
